refactor: 类型文件统一为 types，增加 fixednavitem 类型
